+++ main/src/core/Mono.TextEditor.Platform/Mono.TextEditor.Platform.projitems
@@ -114,7 +114,6 @@
       <Compile Include="$(MSBuildThisFileDirectory)\Platform\Text\Impl\TextModel\Storage\PageManager.cs" />
       <Compile Include="$(MSBuildThisFileDirectory)\Platform\Text\Impl\TextModel\Storage\SimpleTextStorage.cs" />
       <Compile Include="$(MSBuildThisFileDirectory)\Platform\Text\Impl\TextModel\Storage\SimpleTextStorageLoader.cs" />
-      <Compile Include="$(MSBuildThisFileDirectory)\Platform\Text\Impl\TextModel\StringRebuilder\BaseStringRebuilder.cs" />
       <Compile Include="$(MSBuildThisFileDirectory)\Platform\Text\Impl\TextModel\StringRebuilder\BinaryStringRebuilder.cs" />
       <Compile Include="$(MSBuildThisFileDirectory)\Platform\Text\Impl\TextModel\StringRebuilder\StringRebuilder.cs" />
       <Compile Include="$(MSBuildThisFileDirectory)\Platform\Text\Impl\TextModel\StringRebuilder\LineSpan.cs" />
